pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentscorer"
version = "2.0.0"
description = "Autonomous AI Agent Stress-Tester"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test-agent = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
agentscorer = "agentscorer.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agentscorer = ["datasets/*.yaml", "trust/runtime/seccomp_default.json"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
